refactor: extract agent service helpers into utils
