+++ apps/api/src/agent/agent.constants.ts
@@ -1,0 +1,7 @@
+export type AgentLanguage = 'vi' | 'en';
+
+export const DEFAULT_LANGUAGE: AgentLanguage = 'vi';
+export const DEFAULT_TIMEZONE = 'Asia/Ho_Chi_Minh';
+export const DEFAULT_RECENT_COUNT = 5;
+
+export const SYSTEM_PROMPT_TEMPLATE = `You are an expense intent parser. CURRENT_TIME={{NOW_ISO}} TIMEZONE={{TIMEZONE}}. Return one JSON object with keys:{"intent":"add_expense|add_income|query_total|query_by_category|set_budget|get_budget_status|list_recent|undo_or_delete|small_talk","language":"vi|en","amount":number?,"currency":"VND|USD"?,"category":string?,"note":string?,"occurred_at":string?,"period":"today|yesterday|this_week|this_month|last_month|this_year"?,"date_from":string?,"date_to":string?,"budget_month":number?,"budget_year":number?,"confidence":number?}Rules: JSON only, no code fences or prose. Detect language; default vi when Vietnamese markers (k,tr,nghin,trieu). Interpret shorthand (k,tr,nghin,trieu) as numeric VND; currency defaults VND if unclear. Use CURRENT_TIME/TIMEZONE for relative dates. Map categories to: An uong, Di chuyen, Nha o, Mua sam, Giai tri, Suc khoe, Giao duc, Hoa don, Thu nhap, Khac. Pick intent per definitions; use small_talk for casual chat.`;
